inspect: Construct an array of struct iovec using strdup()
set_iovec() did not work well with a string literal because
iov_base whose type was (void *) conflicted with (const char *).
To solve this, strdup() is now always used and the allocated buffers
are released after the nmount(2) call.

In addition to it, handling of md_unit has been fixed.  While it was
declared as unsigned, a negative value could be set to represent
no md(4) device was configured.  It is now long type.